Fix failing CCG tests
